+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - added python3.6 to tests
 - add home_dir config option
+- add better config testing
 
 ### Fixed
 ### Changed